vl-test/requirements.txt
torch>=2.1.0
transformers>=4.52.0
accelerate
peft
Pillow
huggingface_hub
torchvision